+++ mobile/android/android-components/components/support/utils/src/main/java/mozilla/components/support/utils/WebURLFinder.kt
@@ -75,7 +75,8 @@ class WebURLFinder {
     companion object {
         // Taken from mozilla.components.support.ktx.util.URLStringUtils. See documentation
         // there for a complete description.
-        private const val autolinkWebUrlPattern = "(\\w+-)*\\w+(://[/]*|:|\\.)(\\w+-)*\\w+([\\S&&[^\\w-]]\\S*)?"
+        private const val autolinkWebUrlPattern =
+            "(\\w+-+)*[\\w\\[]+(://[/]*|:|\\.)(\\w+-+)*[\\w\\[:]+([\\S&&[^\\w-]]\\S*)?"
 
         private val autolinkWebUrl by lazy {
             Pattern.compile(autolinkWebUrlPattern, 0)